+++ src/features/mypage/hooks/useFetchSalaryData.ts
@@ -3,8 +3,8 @@ import { useDispatch } from 'react-redux';
 import { setAvailableSalaryDates } from '@/redux/salary-slice';
 import { collection, getDocs } from 'firebase/firestore';
 import { auth, db } from '@/firebase';
-import { SalaryData } from '@/pages/mypage/salaryTypes';
-import { formatDate } from '@/pages/mypage/ui/mypage-utils/formatDate';
+import { SalaryData } from '@/features/mypage/types/salaryTypes';
+import { formatDate } from '@/features/mypage/ui/utils/formatDate';
 
 export const useFetchSalaryData = () => {
   const dispatch = useDispatch();
