+++ src/features/mypage/styles/salary-modal.styles.ts
