+++ src/features/mypage/styles/salary-section.styles.ts
