+++ src/features/mypage/types/salaryTypes.ts
