+++ src/features/mypage/ui/utils/formatCurrency.ts
